+++ src/features/seller/resolvers/seller-store-query.resolver.ts
@@ -1,9 +1,20 @@
-import { UseGuards } from '@nestjs/common';
+import { Inject, UseGuards } from '@nestjs/common';
 import { Args, Query, Resolver } from '@nestjs/graphql';
 
 import { SellerCursorInput } from '@/features/seller/dto/inputs/seller-cursor.input';
 import { SellerDateCursorInput } from '@/features/seller/dto/inputs/seller-date-cursor.input';
-import { SellerStoreService } from '@/features/seller/services/seller-store.service';
+import {
+  SELLER_STORE_HOURS_SERVICE,
+  type ISellerStoreHoursService,
+} from '@/features/seller/services/seller-store-hours.service.interface';
+import {
+  SELLER_STORE_POLICY_SERVICE,
+  type ISellerStorePolicyService,
+} from '@/features/seller/services/seller-store-policy.service.interface';
+import {
+  SELLER_STORE_PROFILE_SERVICE,
+  type ISellerStoreProfileService,
+} from '@/features/seller/services/seller-store-profile.service.interface';
 import type {
   SellerCursorConnection,
   SellerStoreBusinessHourOutput,
@@ -21,12 +32,19 @@ import {
 @Resolver('Query')
 @UseGuards(JwtAuthGuard)
 export class SellerStoreQueryResolver {
-  constructor(private readonly storeService: SellerStoreService) {}
+  constructor(
+    @Inject(SELLER_STORE_PROFILE_SERVICE)
+    private readonly profileService: ISellerStoreProfileService,
+    @Inject(SELLER_STORE_HOURS_SERVICE)
+    private readonly hoursService: ISellerStoreHoursService,
+    @Inject(SELLER_STORE_POLICY_SERVICE)
+    private readonly policyService: ISellerStorePolicyService,
+  ) {}
 
   @Query('sellerMyStore')
   sellerMyStore(@CurrentUser() user: JwtUser): Promise<SellerStoreOutput> {
     const accountId = parseAccountId(user);
-    return this.storeService.sellerMyStore(accountId);
+    return this.profileService.sellerMyStore(accountId);
   }
 
   @Query('sellerStoreBusinessHours')
@@ -34,7 +52,7 @@ export class SellerStoreQueryResolver {
     @CurrentUser() user: JwtUser,
   ): Promise<SellerStoreBusinessHourOutput[]> {
     const accountId = parseAccountId(user);
-    return this.storeService.sellerStoreBusinessHours(accountId);
+    return this.hoursService.sellerStoreBusinessHours(accountId);
   }
 
   @Query('sellerStoreSpecialClosures')
@@ -43,7 +61,7 @@ export class SellerStoreQueryResolver {
     @Args('input', { nullable: true }) input?: SellerCursorInput,
   ): Promise<SellerCursorConnection<SellerStoreSpecialClosureOutput>> {
     const accountId = parseAccountId(user);
-    return this.storeService.sellerStoreSpecialClosures(accountId, input);
+    return this.hoursService.sellerStoreSpecialClosures(accountId, input);
   }
 
   @Query('sellerStoreDailyCapacities')
@@ -52,6 +70,6 @@ export class SellerStoreQueryResolver {
     @Args('input', { nullable: true }) input?: SellerDateCursorInput,
   ): Promise<SellerCursorConnection<SellerStoreDailyCapacityOutput>> {
     const accountId = parseAccountId(user);
-    return this.storeService.sellerStoreDailyCapacities(accountId, input);
+    return this.policyService.sellerStoreDailyCapacities(accountId, input);
   }
 }
