+++ src/features/seller/services/seller-store-hours.service.spec.ts
@@ -1,0 +1,282 @@
+import { BadRequestException, NotFoundException } from '@nestjs/common';
+import type { PrismaClient } from '@prisma/client';
+
+import { AUDIT_LOG_REPOSITORY, AuditLogRepository } from '@/features/audit-log';
+import { SellerRepository } from '@/features/seller/repositories/seller.repository';
+import { SellerStoreHoursService } from '@/features/seller/services/seller-store-hours.service';
+import { disconnectTestPrismaClient } from '@/test/db/prisma-test-client';
+import { closeTruncateConnection, truncateAll } from '@/test/db/truncate';
+import { setupSellerWithStore } from '@/test/factories';
+import { createTestingModuleWithRealDb } from '@/test/modules/testing-module.builder';
+
+describe('SellerStoreHoursService (real DB)', () => {
+  let service: SellerStoreHoursService;
+  let prisma: PrismaClient;
+
+  beforeAll(async () => {
+    const { module, prisma: p } = await createTestingModuleWithRealDb({
+      providers: [
+        SellerStoreHoursService,
+        SellerRepository,
+        {
+          provide: AUDIT_LOG_REPOSITORY,
+          useClass: AuditLogRepository,
+        },
+      ],
+    });
+    service = module.get(SellerStoreHoursService);
+    prisma = p;
+  });
+
+  afterAll(async () => {
+    await closeTruncateConnection();
+    await disconnectTestPrismaClient();
+  });
+
+  beforeEach(async () => {
+    await truncateAll();
+  });
+
+  describe('sellerStoreBusinessHours', () => {
+    it('영업시간 row가 없으면 빈 배열', async () => {
+      const { account } = await setupSellerWithStore(prisma);
+      const result = await service.sellerStoreBusinessHours(account.id);
+      expect(result).toEqual([]);
+    });
+
+    it('day_of_week 오름차순으로 영업시간을 반환한다', async () => {
+      const { account, store } = await setupSellerWithStore(prisma);
+      await prisma.storeBusinessHour.createMany({
+        data: [
+          { store_id: store.id, day_of_week: 1, is_closed: false },
+          { store_id: store.id, day_of_week: 0, is_closed: true },
+        ],
+      });
+
+      const result = await service.sellerStoreBusinessHours(account.id);
+
+      expect(result).toHaveLength(2);
+      expect(result[0].dayOfWeek).toBe(0);
+      expect(result[0].isClosed).toBe(true);
+      expect(result[1].dayOfWeek).toBe(1);
+      expect(result[1].isClosed).toBe(false);
+    });
+  });
+
+  describe('sellerStoreSpecialClosures', () => {
+    it('특별 휴무 목록을 cursor 페이지네이션으로 반환한다', async () => {
+      const { account, store } = await setupSellerWithStore(prisma);
+      await prisma.storeSpecialClosure.create({
+        data: {
+          store_id: store.id,
+          closure_date: new Date('2026-04-01'),
+          reason: '정기 휴무',
+        },
+      });
+
+      const result = await service.sellerStoreSpecialClosures(account.id);
+
+      expect(result.items).toHaveLength(1);
+      expect(result.items[0].reason).toBe('정기 휴무');
+      expect(result.nextCursor).toBeNull();
+    });
+
+    it('limit 초과 시 nextCursor를 반환한다', async () => {
+      const { account, store } = await setupSellerWithStore(prisma);
+      for (let i = 1; i <= 3; i++) {
+        await prisma.storeSpecialClosure.create({
+          data: {
+            store_id: store.id,
+            closure_date: new Date(`2026-04-0${i}`),
+          },
+        });
+      }
+
+      const result = await service.sellerStoreSpecialClosures(account.id, {
+        limit: 2,
+      });
+
+      expect(result.items).toHaveLength(2);
+      expect(result.nextCursor).not.toBeNull();
+    });
+
+    it('cursor 페이지네이션이 동작한다', async () => {
+      const { account, store } = await setupSellerWithStore(prisma);
+      for (let i = 0; i < 3; i++) {
+        await prisma.storeSpecialClosure.create({
+          data: {
+            store_id: store.id,
+            closure_date: new Date(`2026-0${i + 1}-01`),
+            reason: `r${i}`,
+          },
+        });
+      }
+      const first = await service.sellerStoreSpecialClosures(account.id, {
+        limit: 2,
+      });
+      expect(first.items).toHaveLength(2);
+      expect(first.nextCursor).not.toBeNull();
+
+      const second = await service.sellerStoreSpecialClosures(account.id, {
+        limit: 2,
+        cursor: first.nextCursor as string,
+      });
+      expect(second.items.length).toBeGreaterThanOrEqual(1);
+    });
+  });
+
+  describe('sellerUpsertStoreBusinessHour', () => {
+    it('dayOfWeek가 0~6 범위 밖이면 BadRequestException', async () => {
+      const { account } = await setupSellerWithStore(prisma);
+      await expect(
+        service.sellerUpsertStoreBusinessHour(account.id, {
+          dayOfWeek: 7,
+          isClosed: false,
+          openTime: new Date('1970-01-01T09:00:00Z'),
+          closeTime: new Date('1970-01-01T18:00:00Z'),
+        }),
+      ).rejects.toThrow(BadRequestException);
+    });
+
+    it('영업일인데 시간이 없으면 BadRequestException', async () => {
+      const { account } = await setupSellerWithStore(prisma);
+      await expect(
+        service.sellerUpsertStoreBusinessHour(account.id, {
+          dayOfWeek: 1,
+          isClosed: false,
+          openTime: null,
+          closeTime: null,
+        }),
+      ).rejects.toThrow(BadRequestException);
+    });
+
+    it('open >= close면 BadRequestException', async () => {
+      const { account } = await setupSellerWithStore(prisma);
+      await expect(
+        service.sellerUpsertStoreBusinessHour(account.id, {
+          dayOfWeek: 1,
+          isClosed: false,
+          openTime: new Date('1970-01-01T18:00:00Z'),
+          closeTime: new Date('1970-01-01T09:00:00Z'),
+        }),
+      ).rejects.toThrow(BadRequestException);
+    });
+
+    it('정상 영업일 upsert는 row를 생성하고 다시 호출하면 같은 row를 갱신한다', async () => {
+      const { account, store } = await setupSellerWithStore(prisma);
+      const created = await service.sellerUpsertStoreBusinessHour(account.id, {
+        dayOfWeek: 1,
+        isClosed: false,
+        openTime: new Date('1970-01-01T09:00:00Z'),
+        closeTime: new Date('1970-01-01T18:00:00Z'),
+      });
+      expect(created.dayOfWeek).toBe(1);
+      expect(created.isClosed).toBe(false);
+
+      const updated = await service.sellerUpsertStoreBusinessHour(account.id, {
+        dayOfWeek: 1,
+        isClosed: false,
+        openTime: new Date('1970-01-01T10:00:00Z'),
+        closeTime: new Date('1970-01-01T20:00:00Z'),
+      });
+      expect(updated.id).toBe(created.id);
+
+      const rows = await prisma.storeBusinessHour.findMany({
+        where: { store_id: store.id },
+      });
+      expect(rows).toHaveLength(1);
+    });
+
+    it('휴무일 upsert는 시간을 null로 저장한다', async () => {
+      const { account } = await setupSellerWithStore(prisma);
+      const result = await service.sellerUpsertStoreBusinessHour(account.id, {
+        dayOfWeek: 0,
+        isClosed: true,
+        openTime: new Date('1970-01-01T09:00:00Z'),
+        closeTime: new Date('1970-01-01T18:00:00Z'),
+      });
+      expect(result.isClosed).toBe(true);
+      expect(result.openTime).toBeNull();
+      expect(result.closeTime).toBeNull();
+    });
+  });
+
+  describe('sellerUpsertStoreSpecialClosure', () => {
+    it('closureId를 줬는데 없으면 NotFoundException', async () => {
+      const { account } = await setupSellerWithStore(prisma);
+      await expect(
+        service.sellerUpsertStoreSpecialClosure(account.id, {
+          closureId: '999999',
+          closureDate: new Date('2026-04-01'),
+          reason: null,
+        }),
+      ).rejects.toThrow(NotFoundException);
+    });
+
+    it('closureId 없으면 신규 생성', async () => {
+      const { account, store } = await setupSellerWithStore(prisma);
+      const result = await service.sellerUpsertStoreSpecialClosure(account.id, {
+        closureDate: new Date('2026-04-01'),
+        reason: '정기 휴무',
+      });
+      expect(result.reason).toBe('정기 휴무');
+
+      const rows = await prisma.storeSpecialClosure.findMany({
+        where: { store_id: store.id },
+      });
+      expect(rows).toHaveLength(1);
+    });
+
+    it('closureId가 있으면 update 경로로 갱신', async () => {
+      const { account, store } = await setupSellerWithStore(prisma);
+      const created = await prisma.storeSpecialClosure.create({
+        data: {
+          store_id: store.id,
+          closure_date: new Date('2026-04-01'),
+          reason: '구 사유',
+        },
+      });
+
+      const result = await service.sellerUpsertStoreSpecialClosure(account.id, {
+        closureId: created.id.toString(),
+        closureDate: new Date('2026-04-02'),
+        reason: '신 사유',
+      });
+
+      expect(result.id).toBe(created.id.toString());
+      expect(result.reason).toBe('신 사유');
+    });
+  });
+
+  describe('sellerDeleteStoreSpecialClosure', () => {
+    it('존재하지 않는 closureId면 NotFoundException', async () => {
+      const { account } = await setupSellerWithStore(prisma);
+      await expect(
+        service.sellerDeleteStoreSpecialClosure(account.id, BigInt(999)),
+      ).rejects.toThrow(NotFoundException);
+    });
+
+    it('soft-delete + audit log 생성', async () => {
+      const { account, store } = await setupSellerWithStore(prisma);
+      const closure = await prisma.storeSpecialClosure.create({
+        data: { store_id: store.id, closure_date: new Date('2026-04-01') },
+      });
+
+      const result = await service.sellerDeleteStoreSpecialClosure(
+        account.id,
+        closure.id,
+      );
+      expect(result).toBe(true);
+
+      const after = await prisma.storeSpecialClosure.findUnique({
+        where: { id: closure.id },
+      });
+      expect(after?.deleted_at).not.toBeNull();
+
+      const auditLogs = await prisma.auditLog.findMany({
+        where: { store_id: store.id, action: 'DELETE' },
+      });
+      expect(auditLogs).toHaveLength(1);
+    });
+  });
+});
